moved the logic of repairing the models away from the handlers. This will make it easier for the view to get variables to display from the model.
