+++ src/hvl/plugin/parmorel/model/ParmorelModelFixer.java
@@ -1,0 +1,46 @@
+package hvl.plugin.parmorel.model;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+
+import hvl.projectparmorel.modelrepair.ModelFixer;
+import hvl.projectparmorel.modelrepair.QModelFixer;
+import hvl.projectparmorel.modelrepair.Solution;
+
+public class ParmorelModelFixer {
+	private ModelFixer modelFixer;
+	private List<Solution> possibleSolutions;
+	
+	public ParmorelModelFixer() {
+		modelFixer = new QModelFixer();
+		possibleSolutions = new ArrayList<Solution>();
+	}
+	
+	/**
+	 * Fixes the selected model with the specified preferences.
+	 * 
+	 * @param preferences
+	 */
+	public List<Solution> fixModel(File model, List<Integer> preferences) {
+		deletePossibleSolutions();
+		
+		modelFixer.setPreferences(preferences);
+		modelFixer.fixModel(model);
+		
+		possibleSolutions =  modelFixer.getPossibleSolutions();
+		return possibleSolutions;
+	}
+
+	
+
+	/**
+	 * Deletes all the files in the possible solutions and clears the list.
+	 */
+	private void deletePossibleSolutions() {
+		for(Solution solution : possibleSolutions) {
+			solution.getModel().delete();
+		}
+		possibleSolutions.clear();
+	}
+}
